DictSettingにsaveToUserDictを追加
